requirements.txt
tensorboard
pillow
torchvision
accelerate==0.34.2
huggingface_hub
transformers==4.45.1
datasets==3.0.1
evaluate==0.4.3
trl==0.11.1
peft==0.13.0
qwen_vl_utils
unsloth